Moved observer files into it's own domain
