+++ src/app/observers/UserObserver.ts
@@ -1,6 +1,6 @@
 import { IUserData } from "@src/core/domains/auth/interfaces/IUserModel";
 import hashPassword from "@src/core/domains/auth/utils/hashPassword";
-import Observer from "@src/core/observer/Observer";
+import Observer from "@src/core/domains/observer/services/Observer";
 
 export default class UserObserver extends Observer<IUserData>
 {
