+++ src/core/base/Model.ts
@@ -5,10 +5,10 @@ import BelongsTo, { BelongsToOptions } from '@src/core/domains/database/mongodb/
 import HasMany, { HasManyOptions } from '@src/core/domains/database/mongodb/relationships/HasMany';
 import { Dates, GetDataOptions, IModel } from '@src/core/interfaces/IModel';
 import IModelData from '@src/core/interfaces/IModelData';
-import { IObserver } from '@src/core/interfaces/observer/IObserver';
-import { WithObserver } from '@src/core/observer/WithObserver';
 import { App } from '@src/core/services/App';
 import Str from '@src/core/util/str/Str';
+import { IObserver } from '../domains/observer/interfaces/IObserver';
+import { WithObserver } from '../domains/observer/services/WithObserver';
 
 export default abstract class Model<Data extends IModelData> extends WithObserver<Data> implements IModel<Data> {
     // The database connection
